refactored after comments in the PR - extracted the classes to PlatformDiscoveryTypes, refactored some methods
